requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
alibabacloud-cdn20180510==1.0.3
alibabacloud-credentials==0.2.0
alibabacloud-alidns20150109==2.0.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.1.5
alibabacloud-slb20140515==2.0.3
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.3.0
alibabacloud-tea-util==0.3.5
alibabacloud-vpc20160428==2.0.2
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-domain==3.14.6
aliyun-python-sdk-ecs==4.24.12
aliyun-python-sdk-kms==2.15.0
aliyun-python-sdk-rds==2.6.1
aliyun-python-sdk-vpc==3.0.16
aliyunsdkcore==1.0.3
amqp==5.0.9
asgiref==3.4.1
astroid==2.8.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.5.7
bcrypt==3.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
click==8.0.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==35.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.7
django-auth-ldap==4.0.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg2==1.19.4
dwebsocket==0.5.12
Flask==2.0.2
frozenlist==1.3.0
grpcio==1.41.0
grpcio-tools==1.41.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
incremental==21.3.0
inflection==0.5.1
isort==5.9.3
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
msgpack==1.0.3
multi-key-dict==2.0.3
multidict==6.0.2
mysqlclient==2.1.0
oss2==2.15.0
packaging==21.3
paramiko==2.8.1
pbr==5.8.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.13.0
PyJWT==1.7.1
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pypandoc==1.6.4
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-jenkins==1.7.0
python-ldap==3.4.0
pytz==2021.1
redis==4.0.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.16.0
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
webshell==0.0.3
websockets==10.1
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
Flask==2.0.2
python-dotenv
requests
pycryptodome
cloudflare==2.8.15
